+++ src/main/java/com/sasd/eventor/controllers/UserController.java
@@ -83,7 +83,12 @@ public class UserController {
     }
 
     @DeleteMapping("/delete")
-    public void deleteById(@RequestParam Long id) {
-        userService.deleteById(id);
+    public void deleteById(@RequestParam Long id, @RequestParam String jwt) {
+        if (userService.findByJwt(jwt).isEmpty()) {
+            throw new EventorException("You need to be authorized");
+        } else {
+            findById(id);
+            userService.deleteById(id);
+        }
     }
 }
